% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caledocensus-package.R
\docType{package}
\name{Caledocensus-package}
\alias{Caledocensus}
\alias{Caledocensus-package}
\title{Caledocensus: Import data from New Caledonian census}
\author{
\strong{Maintainer}: Jonas  Brouillon \email{jonas.brouillon@iac.nc}

}
\keyword{internal}
