% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_pop_communaute.R
\docType{data}
\name{evo_pop_communaute}
\alias{evo_pop_communaute}
\title{Données aggregées des recensements de Nouvelle-Calédonie - Evolution de la population par communauté d'appartenance}
\format{
A tibble with 363 rows and 4variables:
\describe{
\item{Communaute}{}
\item{Province}{}
\item{Annee}{}
\item{Pop}{}
}
}
\source{
\url{https://www.isee.nc/population/recensement/structure-de-la-population-et-evolutions}
}
\usage{
evo_pop_communaute
}
\description{
Jeu de données présentant l'évolution de la population par communauté d'appartenance et province depuis 1989. Chaque ligne représente une communauté, lors d'une année identifiée par la variable "Année".
}
\keyword{datasets}
