% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_pop_communes.R
\docType{data}
\name{evo_pop_communes}
\alias{evo_pop_communes}
\title{Données aggregées des recensements de Nouvelle-Calédonie - Evolution de la population par commune}
\format{
A tibble with 363 rows and 4variables:
\describe{
\item{Province}{}
\item{Commune}{}
\item{Annee}{}
\item{Pop}{}
}
}
\source{
\url{https://www.isee.nc/population/recensement/structure-de-la-population-et-evolutions}
}
\usage{
evo_pop_communes
}
\description{
Jeu de données présentant l'évolution de la population par commune depuis 1969.Chaque ligne représente une commune, lors d'une année identifiée par la variable "Anneée".
}
\keyword{datasets}
