% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_map.R
\docType{data}
\name{nc_map}
\alias{nc_map}
\title{Limites administratives de la Nouvelle- Calédonie et données agrégées du recensement 2019 de Nouvelle-Calédonie - Individus de 18 ans et plus par commune}
\format{
A sf with 33 rows & 38 variables (S3: sf/tbl_df/tbl/data.frame)
\describe{
\item{objectid}
\item{nom}
\item{nom_minus}
\item{code_com}
\item{code_post}
\item{fichier}
\item{Shape__Are}
\item{Shape__Len}
\item{geometry}
\item{Commune}
\item{Province}
\item{Part_femme}{}
\item{Part_homme}{}
\item{Part_europeenne}{}
\item{Part_kanak}{}
\item{Part_wallisienne_futunienne}{}
\item{Part_plusieurs_communautes_metis}{}
\item{Part_autre_et_non_declaree}{}
\item{Part_10_a_19_ans}{}
\item{Part_20_a_29_ans}{}
\item{Part_30_a_39_ans}{}
\item{Part_40_a_49_ans}{}
\item{Part_50_a_59_ans}{}
\item{Part_60_a_69_ans}{}
\item{Part_70_a_79_ans}{}
\item{Part_80_a_89_ans}{}
\item{Part_90_ans_et_plus}{}
\item{Part_agriculteur}{}
\item{Part_artisan_commercant}{}
\item{Part_cadre}{}
\item{Part_profession_intermediaire}{}
\item{Part_employe}{}
\item{Part_ouvrier}{}
\item{Part_retraite}{}
\item{Part_inactif}{}
\item{Pop_18_plus_total}{}
\item{Pop_18_plus_classe}{}
\item{Part_Oui_2020}{}
}
}
\source{
\url{https://data.gouv.nc/explore/dataset/communes-nc-limites-terrestres-simplifiees/information/}
}
\usage{
nc_map
}
\description{
Fichier shp présentant les limites communales de la Nouvelle-Calédonie, ainsi que, pour l'année 2019, les proportions par commune selon le genre, la tranche d'âge, la communauté et la CSP, parmi les individus de 18 ans et plus. La part du Oui (\%) au référendum de 2020 est également inclue dans le jeu de données.
}
\keyword{datasets}
