% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_18plus_communes.R
\docType{data}
\name{pop_18plus_communes}
\alias{pop_18plus_communes}
\title{Données aggregées du recensement 2019 de Nouvelle-Calédonie - Individus de 18 ans et plus par commune}
\format{
A tibble with 33 rows and 26 variables:
\describe{
\item{Province}{}
\item{Commune}{}
\item{Part_femme}{}
\item{Part_homme}{}
\item{Part_europeenne}{}
\item{Part_kanak}{}
\item{Part_wallisienne_futunienne}{}
\item{Part_plusieurs_communautes_metis}{}
\item{Part_autre_et_non_declaree}{}
\item{Part_10_a_19_ans}{}
\item{Part_20_a_29_ans}{}
\item{Part_30_a_39_ans}{}
\item{Part_40_a_49_ans}{}
\item{Part_50_a_59_ans}{}
\item{Part_60_a_69_ans}{}
\item{Part_70_a_79_ans}{}
\item{Part_80_a_89_ans}{}
\item{Part_90_ans_et_plus}{}
\item{Part_agriculteur}{}
\item{Part_artisan_commercant}{}
\item{Part_cadre}{}
\item{Part_profession_intermediaire}{}
\item{Part_employe}{}
\item{Part_ouvrier}{}
\item{Part_retraite}{}
\item{Part_inactif}{}
\item{Pop_18_plus_total}{}
}
}
\source{
\url{https://www.isee.nc/population/recensement/structure-de-la-population-et-evolutions}
}
\usage{
pop_18plus_communes
}
\description{
Chaque ligne représente une commune, identifiée par la variable Commune, et associée à différentes variables mesurant la part parmi les individus de 18 ans et plus.
}
\keyword{datasets}
